+++ argopy/tests/test_data_fetchers.py
@@ -22,7 +22,7 @@ from unittest import TestCase
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
-from argopy.errors import InvalidFetcherAccessPoint, InvalidFetcher, ErddapServerError
+from argopy.errors import InvalidFetcherAccessPoint, InvalidFetcher, ErddapServerError, CacheFileNotFound, FileSystemHasNoCache
 
 from argopy.utilities import list_available_data_src, isconnected, erddap_ds_exists
 AVAILABLE_SOURCES = list_available_data_src()
@@ -139,44 +139,65 @@ class EntryPoints_AllBackends(TestCase):
 class Erddap_backend(TestCase):
     """ Test main API facade for all available dataset of the ERDDAP fetching backend """
 
-    # @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
-    # def test_cachepath(self):
-    #     assert isinstance(ArgoDataFetcher(src='erddap').profile(6902746, 34).fetcher.cachepath, str) == True
+    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+    def test_cachepath_notfound(self):
+        testcachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
+        with argopy.set_options(cachedir=testcachedir):
+            loader = ArgoDataFetcher(src='erddap', cache=True).profile(6902746, 34)
+            with pytest.raises(CacheFileNotFound):
+                loader.fetcher.cachepath
+        shutil.rmtree(testcachedir) # Make sure the cache is empty
+
+    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+    def test_nocache(self):
+        testcachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
+        with argopy.set_options(cachedir=testcachedir):
+            loader = ArgoDataFetcher(src='erddap', cache=False).profile(6902746, 34)
+            ds = loader.to_xarray()
+            with pytest.raises(FileSystemHasNoCache):
+                loader.fetcher.cachepath
 
     @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
     def test_caching_float(self):
-        cachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
-        try:
-            # 1st call to load from erddap and save to cachedir:
-            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).float([1901393, 6902746]).to_xarray()
-            # 2nd call to load from cached file
-            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).float([1901393, 6902746]).to_xarray()
-            assert isinstance(ds, xr.Dataset) == True
-            shutil.rmtree(cachedir)
-        except ErddapServerError: # Test is passed when something goes wrong because of the erddap server, not our fault !
-            shutil.rmtree(cachedir)
-            pass
-        except:
-            shutil.rmtree(cachedir)
-            raise
+        testcachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
+        with argopy.set_options(cachedir=testcachedir):
+            try:
+                loader = ArgoDataFetcher(src='erddap', cache=True).float([1901393, 6902746])
+                # 1st call to load from erddap and save to cachedir:
+                ds = loader.to_xarray()
+                # 2nd call to load from cached file:
+                ds = loader.to_xarray()
+                assert isinstance(ds, xr.Dataset) == True
+                assert isinstance(loader.fetcher.cachepath, str) == True
+                shutil.rmtree(testcachedir)
+            except ErddapServerError: # Test is passed when something goes wrong because of the erddap server, not our fault !
+                shutil.rmtree(testcachedir)
+                pass
+            except:
+                shutil.rmtree(testcachedir)
+                raise
 
     @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
     def test_caching_profile(self):
-        cachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
-        try:
-            # 1st call to load from erddap and save to cachedir:
-            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
-            # 2nd call to load from cached file
-            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
-            assert isinstance(ds, xr.Dataset) == True
-            shutil.rmtree(cachedir)
-        except ErddapServerError: # Test is passed when something goes wrong because of the erddap server, not our fault !
-            shutil.rmtree(cachedir)
-            pass
-        except:
-            shutil.rmtree(cachedir)
-            raise
+        testcachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
+        with argopy.set_options(cachedir=testcachedir):
+            loader = ArgoDataFetcher(src='erddap', cache=True).profile(6902746, 34)
+            try:
+                # 1st call to load from erddap and save to cachedir:
+                ds = loader.to_xarray()
+                # 2nd call to load from cached file
+                ds = loader.to_xarray()
+                assert isinstance(ds, xr.Dataset) == True
+                assert isinstance(loader.fetcher.cachepath, str) == True
+                shutil.rmtree(testcachedir)
+            except ErddapServerError: # Test is passed when something goes wrong because of the erddap server, not our fault !
+                shutil.rmtree(testcachedir)
+                pass
+            except:
+                shutil.rmtree(testcachedir)
+                raise
 
+    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
     def test_N_POINTS(self):
         n = ArgoDataFetcher(src='erddap').region([-70, -65, 35., 40., 0, 10., '2012-01', '2013-12']).fetcher.N_POINTS
         assert isinstance(n, int) == True