+++ docs/whats-new.rst
@@ -23,6 +23,8 @@ v0.1.4 (XX June 2020)
 
 - Now uses `fsspec <https://filesystem-spec.readthedocs.io>`_ as file system for caching local files (:pr:`19`). This closes issues :issue:`12`, :issue:`15` and :issue:`17`. By `G. Maze <http://www.github.com/gmaze>`_.
 
+- Erddap fetcher now uses netcdf format to retrieve data (:pr:`19`).
+
 v0.1.3 (15 May 2020)
 --------------------